\name{auto}
\alias{auto}
\docType{data}
\title{Run off triangle of accumulated claim data}
\description{Run-off triangles of Personal Auto and Commercial Auto insurance.}
  
\usage{data(auto)}
\format{
  A list of three matrices, paid Personal Auto, incurred Personal Auto and paid Commercial Auto respectively.
}
\source{
	\cite{Zhang (2010). A general multivariate chain ladder model. \emph{Insurance: Mathematics and Economics}, 46, pp. 588-599.}
}
\examples{
data(auto)
names(auto)
}
\keyword{datasets}

