% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{BootMackChainLadder}
\alias{BootMackChainLadder}
\title{Boot Mack Chain Ladder model}
\usage{
BootMackChainLadder(
  Triangle,
  B = 100,
  distNy = "normal",
  threshold = NA,
  BF.premiums = NULL,
  BF.param = c(5, 5),
  stab = NA,
  clusters = NA
)
}
\arguments{
\item{Triangle}{A simple triangle containing the data.}

\item{B}{The number of bootstrap resamples}

\item{distNy}{Distribution of next-year incremental payments. Either "normal" or "residuals", see details}

\item{threshold}{A value of NA (default) will prevent exclusion of residuals, but a numeric value (e.g 2) will exclude all residuals that have an absolution value greater than 2.}

\item{BF.premiums}{If a Bornhuetter-fergusson is needed, input a vector of Ultimates Premiums here. Otherwise, the BF code will not be triggered.}

\item{BF.param}{A vector of 2 integers that represent (respectively) the number of year of averaging Loss-ratios for the Bornhuetter Fergusson and then the number of year of applying the Bornhuetter Fergusson to.}

\item{stab}{A stabilisation parameter: forces developpements factors corresponding to developpements years greater than `stab` to be one.}

\item{clusters}{list of sequences of column numbers. See details.}
}
\value{
A BootMackChainLadder object with a lot of information about the bootstrapping. You can `plot` it, `print` it and `str` it to extract information. Functions `mean`, `CDR` and `summary` can be used to extract informations as well.
}
\description{
This function implement a simple bootstrap of the residuals from Mack model, potentially including a one-year risk view.
}
\details{
The bootstrap that is implemented here consist in a resampling of residuals obtained by the Mack model
(or simulated standard normal residuals if you choose so through the `distNy` parameter). On those resampled triangles, we construct a one-year
point of view of the Mack model, allowing us to bootstrap one-year quantities like the CDR or next year IBNRS.
Using this function properly, you could check that the proposed bootstrap is convergent with the Merz-Wuthrich formula
if you take standard normal residuals, but not otherwise.

The residuals can also be clustered through the `clusters` argument, by providing clusters as a list of vectors of integers corresponding to each cluster:
`clusters=list(seq(1,10),seq(11,15))` for a triangle with 15 columns will cluster the residuals in two parts. The meaning of this parameter is that the residuals are resampled inside clusters only:
the underlying i.i.d hypothesis that is needed for the resampling can be tampered with through this parameterisation

For more details about the model, see the vignette: 
\code{vignette("BootMack", package = "ChainLadder")}
}
\examples{
data(ABC)
BootMackChainLadder(Triangle = ABC, B = 100, distNy = "residuals", threshold = 2)
}
