% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{CDR.BootMackChainLadder}
\alias{CDR.BootMackChainLadder}
\title{CDR}
\usage{
\method{CDR}{BootMackChainLadder}(x, ...)
}
\arguments{
\item{x}{A BootMackChainLadder Object}

\item{...}{other arguments for other methods.}
}
\value{
A data.Frame with one-year results from the bootstrap.
}
\description{
Calculate the one-year Claim developpement results from the bootstrapped mack model.
}
\examples{
data(ABC)
BMCL <- BootMackChainLadder(Triangle = ABC, B = 100, distNy = "residuals", threshold = 2)
CDR(BMCL)
}
