\name{liab}
\alias{liab}
%\docType{liab}
\title{Run off triangle of accumulated claim data}
\description{Run-off triangles of General Liability and Auto Liability.}
  
\usage{data(auto)}
\format{
  A list of two matrices, General Liability and Auto Liability respectively.
}
\source{
	\cite{Braun C (2004). The prediction error of the chain ladder method applied to correlated run off triangles. \emph{ASTIN Bulletin} 34(2): 399-423}
}
\examples{
data(liab)
names(liab)
}
\keyword{datasets}

