% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{mean.BootMackChainLadder}
\alias{mean.BootMackChainLadder}
\title{mean.BootMackChainLadder}
\usage{
\method{mean}{BootMackChainLadder}(x, ...)
}
\arguments{
\item{x}{A BootMackChainLadder Object}
}
\value{
Three data.frames containing data per origin year ("ByOrigin"), by developpement year ("ByDev) and globaly ("Totals)
}
\description{
Calculate mean statiscics from the bootstraped mack model
}
\examples{
data(ABC)
BMCL <- BootMackChainLadder(Triangle = ABC, B = 100, distNy = "residuals", threshold = 2)
mean(BMCL)
}
