% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{MultiBootMackChainLadder}
\alias{MultiBootMackChainLadder}
\title{MultiBootMackChainLadder}
\usage{
MultiBootMackChainLadder(
  triangles,
  B = 100,
  distNy = "normal",
  names = NULL,
  threshold = NA,
  stab = NA
)
}
\arguments{
\item{triangles}{A List of Triangles objects of the same size.}

\item{B}{The number of boostrap replicates}

\item{distNy}{The process distribution for next year increments. Either "normal" (default), "residuals.bycolumn","residuals.global" or "residuals". See details below.}

\item{names}{enventual names of the different triangles. IF set to NULL, the procedure will try to get names from the triangles list (first argument)}

\item{threshold}{Eventual exclusions limit for residuals. Set to NA (default) to avoid excluding anything.}
}
\value{
a MBMCL object containing a list of BMCL objects and a little more.
}
\description{
The multi boot mack chain ladder algorythme computes todays and next-year common estimates on a portefolio of several triangles, following closely a synchronised version of BootMackChainLadder.
}
\details{
This function calculates, on a list of triangles, a synchronised bootstrap in the Mack model for the triangles. 
The One-year point of view is also calculated. 
It returns a specific S3 object, on wich you can use the functionsmean, CDR, NyIBNR, Corel provided by the package. 

This model uses the fact that the Mack model can be seen as a quasi-glm, and therefore provide resiuals.
Bootstrapping thoose residuals on the upper-left triangle allows to get bootstrap distribution of today's estimations (reserves, ultimates, ...).
Furthermore, if you simulate net year payments with a given process ditribution in each simulation, it gives 1 year results.


If distNy = "normal", it follows *boumezoued & al* and converges to the Merz-Wüthrich formula in the Braun model. 
If distNy = "residuals", it also converges strongly but NOT to the Merz-Wûthrich formula. Same for ohter possibilities. 
Notes that if you choose distNy = "residuals.bycolumn", the residuals will be resampled inside each column and not accross columns. Setting distNy = "residuals.global" or "residuals" triggers the same code.
}
\examples{
data(ABC)
triangles <- list(tri1 = ABC, tri2 = ABC, tri3 = ABC)
MultiBootmackChainLadder(triangles,100)
}
\seealso{
BootMackChainLadder
}
