\name{MW2008}
\alias{MW2008}
\encoding{UTF-8}
\docType{data}
\title{Run-off claims triangle}
\description{
Cumulative claims development triangle
}
\format{
  A matrix with 9 accident years and 9 development years.
}
\source{
\cite{Modelling the claims development result for solvency purposes.
Michael Merz, Mario V. Wüthrich. Casualty Actuarial Society E-Forum, Fall 2008.}
}

\examples{
MW2008
plot(MW2008, lattice=TRUE)
}
\keyword{datasets}