\name{MW2014}
\alias{MW2014}
\encoding{UTF-8}
\docType{data}
\title{Run-off claims triangle}
\description{
Cumulative claims development triangle
}
\format{
  A matrix with 17 accident years and 17 development years.
}
\source{
\cite{Claims Run-Off Uncertainty: The Full Picture. Michael Merz, 
Mario V. Wüthrich. Swiss Finance Institute Research Paper No. 14-69. 2014}
}
\examples{
MW2014
plot(MW2014, lattice=TRUE)
}
\keyword{datasets}