\name{MWR}
\alias{MWR}
\docType{data}
\title{Run-off claims triangle}
\description{
Cumulative claims development triangle
}
\format{
  A matrix with 9 accident years and 9 development years.
}
\source{
\cite{Modelling the claims development result for solvency purposes.
Michael Merz, Mario V. Wuethrich. Casualty Actuarial Society E-Forum, Fall 2008.}
}
\examples{
MWR
plot(MWR, lattice=TRUE)
}
\keyword{datasets}