\name{NullNum-class}
\docType{class}
\alias{NullNum-class}
\alias{NullChar-class}

\title{Class "NullNum" and "NULLChar }
\description{	Virtual class for \code{c("null", "numeric")} and \code{c("null","character"}}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "NullNum" in the signature.
}

\keyword{classes}
