% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{NyIBNR}
\alias{NyIBNR}
\title{Title}
\usage{
NyIBNR(x, ByOrigin = FALSE)
}
\arguments{
\item{x}{A MultiBootMackChainLadder Object}

\item{ByOrigin}{If Set to TRUE, Next year IBNRS will be rgiveng with per-origin details.}
}
\value{
Data frame with next year informations.
}
\description{
Title
}
\details{
Gives a data.frame with next-year payments informations found by the MBMCL model.
}
\examples{
data(ABC)
triangles <- list(tri1 = ABC, tri2 = ABC, tri3 = ABC)
MBMCL <- MultiBootmackChainLadder(triangles,100)
NyIBNR(MBMCL)
}
