% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PIC.R
\name{PaidIncurredChain}
\alias{PaidIncurredChain}
\title{PaidIncurredChain}
\usage{
PaidIncurredChain(triangleP, triangleI)
}
\arguments{
\item{triangleP}{Please explain}

\item{triangleI}{Please explain}
}
\value{
a numeric vector of random deviates of length \code{100 * x}
}
\description{
This is the Description section
}
\details{
This is the Details section
}
\examples{
PaidIncurredChain(USAApaid, USAAincurred)
}
\author{
your name
}
\seealso{
\code{\link{MackChainLadder}}
}

