% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{print.BootMackChainLadder}
\alias{print.BootMackChainLadder}
\title{print}
\usage{
\method{print}{BootMackChainLadder}(x, ...)
}
\arguments{
\item{x}{the BootMackChainLadder object}

\item{...}{nothing}
}
\value{
the BMCL object
}
\description{
print
}
\examples{
data(ABC)
BMCL <- BootMackChainLadder(Triangle = ABC, B = 100, distNy = "normal", threshold = 2)
print(BMCL)
}
