\name{RC_report}
\alias{RC_report}
\title{
Ale please provide a title here
}
\description{
A concise (1-5 lines) description of what the function does.
}
\usage{
RC_report(res, q = c(0.4, 0.5, 0.7, 0.9, 0.95, 0.995, 0.9993), RC.var = 0.995)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res}{
    Describe \code{res} here
}
  \item{q}{
     Describe \code{q} here
}
  \item{RC.var}{
     Describe \code{RC.var} here
}
}
\details{
 If necessary, more details than the description above 
}
\value{
 Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}

}
\references{
 Put references to the literature/web site here
}
\author{
 Who to complain to
}
\note{
Further notes
}

\section{Warning }{Make other sections like Warning}

\seealso{
 See Also as \code{\link{help}}
}
\examples{
## Example code here
}
\keyword{ print }
