% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{summary.BootMackChainLadder}
\alias{summary.BootMackChainLadder}
\title{summary.BootMackChainLadder}
\usage{
summary.BootMackChainLadder(object, ...)
}
\arguments{
\item{object}{A BootMackChainLadder Object}
}
\description{
Give summary statistics about the bootstrap.
}
\details{
The function only print information about the model.
}
\examples{
data(ABC)
BMCL <- BootMackChainLadder(Triangle = ABC, B = 100, distNy = "residuals", threshold = 2)
summary(BMCL)
}
