\name{USAA triangle}
\alias{USAApaid, USAAincurred}
\docType{data}
\title{
Example paid and incurred triangle data from
}
\description{
A concise (1-5 lines) description of the dataset. 
}
\usage{data("USAApaid")}
\format{
  The format is:
 triangle [1:10, 1:10] 271778 296828 334331 354530 423048 ...
 - attr(*, "dimnames")=List of 2
  ..$ origin: chr [1:10] "0" "1" "2" "3" ...
  ..$ dev   : chr [1:10] "1" "2" "3" "4" ...
 - attr(*, "class")= chr [1:2] "triangle" "matrix"
}
\details{
Please write a bit more
}
\source{
CAS website
}
\references{
CAS
}
\examples{
data(USAApaid)
## maybe str(USAApaid) ; plot(USAApaid) ...
}
\keyword{datasets}
