% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootMackChainLadder.R
\name{CDR.MultiBootMackChainLadder}
\alias{CDR.MultiBootMackChainLadder}
\title{CDR.MultiBootMackChainLadder}
\usage{
\method{CDR}{MultiBootMackChainLadder}(x)
}
\arguments{
\item{x}{A MultiBootMackChainLadder Object}
}
\value{
Informations about CDR and IBNRS.
}
\description{
CDR.MultiBootMackChainLadder
}
\details{
Give information about next year volatility for the MBMCL model.
}
\examples{
data(ABC)
triangles <- list(tri1 = ABC, tri2 = ABC, tri3 = ABC)
MBMCL <- MultiBootmackChainLadder(triangles,100)
CDR(MBMCL)
}
\seealso{
MultiBootMackChainLadder, CDR.BootMackChainLadder
}
