\name{NullNum-class}
\docType{class}
\alias{NullNum-class}
\alias{NullChar-class}
\alias{NullList-class}

\title{Class "NullNum", "NullChar" and "NullList" }
\description{	Virtual class for \code{c("null", "numeric")}, \code{c("null","character"} and  \code{c("null","list"}}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "NullNum" in the signature.
}

\keyword{classes}
