% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PIC.R
\name{PaidIncurredChain}
\alias{PaidIncurredChain}
\title{PaidIncurredChain}
\usage{
PaidIncurredChain(triangleP, triangleI)
}
\arguments{
\item{triangleP}{Cumulative claims payments triangle}

\item{triangleI}{Incurred losses triangle.}
}
\value{
The function returns:
\describe{
  \item{Res.Origin}{Claims reserves for different origin years.}
  \item{Res.Tot}{Total reserve.}
  \item{s.e.}{Square root of mean square error of prediction
  for the total reserve.}
}
}
\description{
The Paid-incurred Chain model (Merz, Wuthrich (2010)) combines
claims payments and incurred losses information to get
a unified ultimate loss prediction.
}
\details{
The method uses some basic properties of multivariate Gaussian distributions
to obtain a mathematically rigorous and consistent model for the combination
of the two information channels.

Perhaps some details here
}
\note{
The model is implemented in the special case of non-informative priors.
}
\examples{
PaidIncurredChain(USAApaid, USAAincurred)
}
\author{
Fabio Concina, \email{fabio.concina@gmail.com}
}
\references{
Merz, M., Wüthrich, M. (2010). Paid–incurred chain claims reserving method.
Insurance: Mathematics and Economics, 46(3), 568–579.
}
\seealso{
\code{\link{MackChainLadder}},\code{\link{MunichChainLadder}}
}

