\name{residCov}
\docType{genericFunction}
\alias{residCov}
\alias{residCov-methods}
\alias{residCov,MultiChainLadder-method}
\title{ Generic function for residCov}
\description{
 \code{residCov}  is a generic function to extract residual covariance from a system of fitted regressions. 
}

\usage{
residCov(object,...)

\S4method{residCov}{MultiChainLadder}(object,...)
}

\arguments{
  \item{object}{An object of class "MultiChainLadder". }
  \item{\dots}{ Currently not used. }
  }

\author{ Wayne (Yanwei) Zhang \email{actuaryzhang@uchicago.edu} }

\seealso{See also  \code{\linkS4class{MultiChainLadder}}. }
\keyword{methods}

